extended hal to 32bit, added mmcsd settings

+++ src/system/hal/common/hal.c
@@ -7,20 +7,20 @@
 
 #include "hal.h"
 
-inline mmio_t hal_get_register(uint32_t port, uint8_t offset)
+inline mmio_t hal_get_register(uint32_t port, uint32_t offset)
 {
 	return (mmio_t)(port + offset);
 }
 
 
-uint32_t hal_get_address_value(uint32_t port, uint8_t offset)
+uint32_t hal_get_address_value(uint32_t port, uint32_t offset)
 {
 	mmio_t reg = hal_get_register(port, offset);
 	return *reg;
 }
 
 
-uint32_t hal_bitmask_write(uint32_t port, uint8_t offset, uint32_t mask, uint8_t size)
+uint32_t hal_bitmask_write(uint32_t port, uint32_t offset, uint32_t mask, uint32_t size)
 {
 	mmio_t reg = hal_get_register(port, offset);
 	uint32_t value = *reg;
@@ -42,14 +42,14 @@ uint32_t hal_bitmask_write(uint32_t port, uint8_t offset, uint32_t mask, uint8_t
 	return value;
 }
 
-void hal_bitmask_set(uint32_t port, uint8_t offset, uint32_t mask)
+void hal_bitmask_set(uint32_t port, uint32_t offset, uint32_t mask)
 {
 	mmio_t reg = hal_get_register(port, offset);
 	*reg |= mask;
 }
 
 
-void hal_bitmask_clear(uint32_t port, uint8_t offset, uint32_t mask)
+void hal_bitmask_clear(uint32_t port, uint32_t offset, uint32_t mask)
 {
 	mmio_t reg = hal_get_register(port, offset);
 	*reg &= ~mask;
